pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dexterous-hand"
version = "0.1.0"
description = "Train a 24-DOF Shadow Hand for contact-rich manipulation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.2.3",
    "imageio[ffmpeg]>=2.37.3",
    "mujoco>=3.6.0",
    "stable-baselines3[extra]>=2.7.1",
    "torch",
    "wandb>=0.25.1",
]

[project.optional-dependencies]
# RunPod GPU MJX + SBX training. Not needed for local PC or RunPod CPU MuJoCo.
mjx = [
    "mujoco-mjx>=3.7.0",
    "sbx-rl>=0.25.0",
    "jax[cuda12]>=0.4.26",
    "flax>=0.9.0",
    "optax>=0.2.3",
    "chex>=0.1.86",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
name = "pytorch-cpu"
explicit = true

# Windows pulls CPU torch from the pytorch-cpu index; Linux (RunPod) gets
# default CUDA torch from PyPI. Same `uv sync` command works on both.
[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "mypy>=1.14",
    "pytest>=8.0",
    "ruff>=0.11",
    "vulture>=2.13",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["dexterous_hand"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["mujoco.*", "stable_baselines3.*", "wandb.*", "gymnasium.*", "imageio.*", "jax.*", "jaxlib.*", "flax.*", "optax.*", "chex.*", "sbx.*", "mujoco.mjx.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scripts.*"]
disallow_untyped_defs = false
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = ["slow: tests requiring MuJoCo environments"]

[tool.vulture]
paths = ["dexterous_hand", "scripts", "main.py"]
min_confidence = 70
ignore_decorators = ["@pytest.fixture", "@abstractmethod"]
exclude = ["tests/"]
